# openapi_hf_space.yaml
openapi: 3.0.3
info:
  title: Cryptocurrency Data Source API - HuggingFace Space
  version: 1.0.0
  description: |
    # مستندات API منبع داده ارز دیجیتال
    
    این API به عنوان **ارائه‌دهنده واحد** برای تمام داده‌های cryptocurrency عمل می‌کند.
    
    ## اولویت‌ها (Priority)
    1. **HF-first (HTTP)** - ابتدا از endpoints داخلی HuggingFace استفاده می‌شود
    2. **WS-only exception** - WebSocket فقط برای endpoint‌های خاص
    3. **Fallback-last** - در صورت عدم موفقیت از providers پشتیبان استفاده می‌شود
    
    ## منابع داده (Data Sources)
    - **Primary**: HuggingFace Space endpoints
    - **Fallback**: External providers (CoinGecko, Binance, WhaleAlert, etc.)
    - **Config**: `/mnt/data/api-config-complete.txt`
    
    ## Meta Fields
    تمام response‌ها شامل فیلد `meta` با اطلاعات زیر:
    - `source`: منبع داده (hf, hf-ws, یا URL provider)
    - `cache_ttl_seconds`: مدت زمان cache
    - `generated_at`: زمان تولید (ISO 8601)
    - `attempted`: لیست منابع تلاش شده (در صورت خطا)
    
  contact:
    name: Amin - Crypto Data API
    url: https://really-amin-datasourceforcryptocurrency.hf.space
  license:
    name: MIT
servers:
  - url: https://really-amin-datasourceforcryptocurrency.hf.space
    description: Production HuggingFace Space
  - url: http://localhost:7860
    description: Local Development
...
paths:
...
  /api/news:
    get:
      tags: [News]
      summary: دریافت اخبار
      description: لیست اخبار cryptocurrency
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: source
          in: query
          schema:
            type: string
          description: فیلتر بر اساس منبع خبر
          example: CoinDesk
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsResponse'
  /api/news/{id}:
    get:
      tags: [News]
      summary: دریافت یک خبر خاص
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsArticle'
        '404':
          description: خبر یافت نشد
...
  /api/crypto/whales/transactions:
    get:
      tags: [Whale Tracking]
      summary: تراکنش‌های نهنگ‌ها
      description: |
        ردیابی تراکنش‌های بزرگ cryptocurrency.
        **Fallback**: اگر HF فراهم نکند، از WhaleAlert/BitQuery استفاده می‌شود
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: chain
          in: query
          schema:
            type: string
            enum: [ethereum, bitcoin, tron, bsc, all]
            default: all
          description: blockchain مورد نظر
        - name: min_amount_usd
          in: query
          schema:
            type: number
            default: 1000000
          description: حداقل مبلغ به دلار
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WhaleTransaction'
                  meta:
                    $ref: '#/components/schemas/MetaInfo'
  /api/crypto/whales/stats:
    get:
      tags: [Whale Tracking]
      summary: آمار نهنگ‌ها
      description: آمار جمع‌آوری شده از تراکنش‌های نهنگ‌ها
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
          description: بازه زمانی (ساعت)
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhaleStatsResponse'
  
  # ============================================================================
  # BLOCKCHAIN ENDPOINTS
  # ============================================================================
  /api/crypto/blockchain/gas:
    get:
      tags: [Blockchain]
      summary: قیمت Gas
      description: هزینه‌های gas برای blockchain‌های مختلف
      parameters:
        - name: chain
          in: query
          required: true
          schema:
            type: string
            enum: [ethereum, bsc, polygon]
          example: ethereum
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GasResponse'
  /api/crypto/blockchain/stats:
    get:
      tags: [Blockchain]
      summary: آمار blockchain
      description: آمار عمومی یک blockchain
      parameters:
        - name: chain
          in: query
          required: true
          schema:
            type: string
            enum: [ethereum, bitcoin, bsc, tron]
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
          description: بازه زمانی
      responses:
        '200':
          description: موفقیت‌آمیز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainStats'
  
  # ============================================================================
  # PROVIDERS & SYSTEM ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key برای endpoint‌های محافظت شده
  schemas:
...
    MetaInfo:
      type: object
      required:
        - source
        - generated_at
      properties:
        source:
          type: string
          description: منبع داده (hf, hf-ws, یا URL provider)
          example: hf
        cache_ttl_seconds:
          type: integer
          description: مدت زمان cache (ثانیه)
          example: 30
        generated_at:
          type: string
          format: date-time
          description: زمان تولید response (ISO 8601)
        attempted:
          type: array
          items:
            type: string
          description: لیست منابع تلاش شده (فقط در صورت خطا)
          example: ["hf", "coingecko", "binance"]
          
    # Market Data Schemas
...
    NewsArticle:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
          format: uri
        summary:
          type: string
        source:
          type: string
        published_at:
          type: string
          format: date-time
        sentiment:
          type: object
          properties:
            label:
              type: string
              enum: [positive, negative, neutral]
            score:
              type: number
              format: float
    NewsResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/NewsArticle'
        total:
          type: integer
        meta:
          $ref: '#/components/schemas/MetaInfo'
...
    WhaleTransaction:
      type: object
      properties:
        id:
          type: string
        tx_hash:
          type: string
          description: Transaction hash
        chain:
          type: string
          enum: [ethereum, bitcoin, tron, bsc]
        from:
          type: string
          description: آدرس مبدا
        to:
          type: string
          description: آدرس مقصد
        amount_usd:
          type: number
          format: float
          description: مبلغ به دلار
        token:
          type: string
          description: نام توکن
        block:
          type: integer
          description: شماره بلاک
        tx_at:
          type: string
          format: date-time
          description: زمان تراکنش
    WhaleStatsResponse:
      type: object
      properties:
        period_hours:
          type: integer
        total_transactions:
          type: integer
        total_volume_usd:
          type: number
          format: float
        top_tokens:
          type: array
          items:
            type: object
            properties:
              token:
                type: string
              count:
                type: integer
              volume_usd:
                type: number
        meta:
          $ref: '#/components/schemas/MetaInfo'
    
    # Blockchain Schemas
    GasResponse:
      type: object
      properties:
        chain:
          type: string
        fast:
          type: number
          format: float
          description: سریع (Gwei یا واحد مربوطه)
        standard:
          type: number
          format: float
        slow:
          type: number
          format: float
        unit:
          type: string
          example: Gwei
        meta:
          $ref: '#/components/schemas/MetaInfo'
    BlockchainStats:
      type: object
      properties:
        chain:
          type: string
        blocks:
          type: integer
          description: تعداد بلاک‌های تولید شده
        txs:
          type: integer
          description: تعداد تراکنش‌ها
        avg_gas:
          type: number
          description: میانگین gas
        pending:
          type: integer
          description: تراکنش‌های در انتظار
        period_hours:
          type: integer
        meta:
          $ref: '#/components/schemas/MetaInfo'
    
    # Provider Schema
...